Refactor the models package directory to group related things better together
Also note that there is a breaking change which has been refactored as part of this - UserProfileData has been split into UserProfileInsertData and UserProfileUpdateData. As of now they are the same but the split helps with clarity and may make future refactors easier.
