+++ apps/back-end/src/server/routes/currentUser/putCurrentUserProfile.ts
@@ -1,8 +1,6 @@
-import type { UserProfileData } from "@lexicon/models";
 import type { Router } from "express";
-import type { ParamsDictionary } from "express-serve-static-core";
 
-import { parseUser } from "@lexicon/models";
+import { parseUser, parseUserProfileUpdateData } from "@lexicon/models";
 
 import { getConnection } from "src/database/connection";
 import editUserProfile from "src/services/users/editUserProfile";
@@ -13,14 +11,15 @@ function putCurrentUserProfile(currentUser: Router) {
   currentUser.put(
     "/profile",
     requireAuth,
-    handleEndpointMiddleware<ParamsDictionary, unknown, UserProfileData>(
-      async (request, response) => {
-        const connection = getConnection();
-        const user = parseUser(request.user);
-        await editUserProfile(connection, user.id, request.body);
-        response.status(200).send({});
-      },
-    ),
+    handleEndpointMiddleware(async (request, response) => {
+      const connection = getConnection();
+      const user = parseUser(request.user);
+
+      const data = parseUserProfileUpdateData(request.body);
+
+      await editUserProfile(connection, user.id, data);
+      response.status(200).send({});
+    }),
   );
 }
