+++ packages/models/src/userSessions/index.ts
@@ -1,0 +1,5 @@
+export { parseCreateUserSessionData } from "src/userSessions/CreateUserSessionData";
+export { parseUserSession } from "src/userSessions/UserSession";
+
+export type { CreateUserSessionData } from "src/userSessions/CreateUserSessionData";
+export type { UserSession } from "src/userSessions/UserSession";
